refactor(sources): extract shared host-safety into Source CDK (#956)
Move the numeric-IPv4-aware host/IP safety classification out of the SentinelOne source into a new typed internal/sourcecdk helper (HostIsUnsafe/HostIsLoopback). Pure relocation with no change to emitted events, URNs, cursors, or validation behavior; lowers the sentinelone LOC ratchet 2181->2089.
