+++ docs/engineering/source-cdk-extraction.md
@@ -32,5 +32,5 @@ Legacy sources above the 300 LOC budget are grandfathered with exact no-growth c
 | `grc` | 1378 | Keep control/evidence shaping out of source orchestration and push common mapping into shared helpers. |
 | `okta` | 2361 | Extract client, pagination, and identity/group/application readers into smaller units. |
 | `panopticon` | 820 | Separate request plumbing from emitted record construction. |
-| `sentinelone` | 2181 | Extract API paging, agent/application readers, and shared response normalization. |
+| `sentinelone` | 2089 | Extract API paging, agent/application readers, and shared response normalization. |
 | `vulnview` | 1064 | Split feed/client access from vulnerability record normalization. |
